Start Modelling Experiment Infrastructure
Add prediction methods, add method for generating bleu scores,
rework/refactor certain bits of the code,

Add an experiments section, with comments on some of the desired
experiments to run, as well as the methods for performing the transfer
training in multiple different ways.

Start using more data, all datasets can be found on the link provided in
the readme, filenames were renamed for my own ease of reading.

Still a lot of work needs to be done. I realized I hadn't commited in a
while, and that the commit was about to be gigantic...

Signed-off-by: James Bodzay <james.bodzay@mail.mcgill.ca>